Updates `copyAndAdvance` in the quote parsing helpers to use an iterator-distance-appropriate type when advancing iterators, improving portability across platforms/ABIs where `long` may not match the iterator difference type
Replaced `static_cast<long>(amount)` with `static_cast<std::ptrdiff_t>(amount)` for the `std::next` offset in `copyAndAdvance`.

---------

+++ Src/AttestationLibrary/src/QuoteVerification/QuoteParsers.h
@@ -45,7 +45,7 @@ inline bool copyAndAdvance(T &val, std::vector<uint8_t>::const_iterator &from, s
     if (available < 0 || (unsigned) available < amount) {
         return false;
     }
-    const auto end = std::next(from, static_cast<long>(amount));
+    const auto end = std::next(from, static_cast<std::ptrdiff_t>(amount));
     return val.insert(from, end);
 }
